Update voter reset command

+++ src/Console/Command/VoterResetCommand.php
@@ -13,11 +13,12 @@ use Doctrine\ORM\ORMException;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
- * Class VotixDropVotersCommand
+ * Class VoterResetCommand
  */
-class DropVotersCommand extends Command
+class VoterResetCommand extends Command
 {
     /**
      * @var VoterRepository
@@ -33,7 +34,7 @@ class DropVotersCommand extends Command
     protected function configure(): void
     {
         $this
-            ->setName('votix:drop')
+            ->setName('votix:voter:reset')
             ->setDescription('Drop voters')
         ;
     }
@@ -48,7 +49,15 @@ class DropVotersCommand extends Command
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $this->voterRepository->deleteAll();
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('Are you sure you want to drop all voters ?', false);
+
+        if ($helper->ask($input, $output, $question)) {
+            $this->voterRepository->deleteAll();
+            $output->writeln('done');
+        } else {
+            $output->writeln('cancelled');
+        }
 
         return Command::SUCCESS;
     }
